+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Core] Fix `Method::canAccess` is not available on Android ([#3249](https://github.com/cucumber/cucumber-jvm/issues/3249))
 
 ## [8.0.1] - 2026-09-24
 ### Fixed